split archive into seperated files
