+++ src/CMakeLists.txt
@@ -98,6 +98,9 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/types.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/verify.c"
         "${CMAKE_CURRENT_LIST_DIR}/serialized-utilities.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/frame.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/depth-sensor.h"
+        "${CMAKE_CURRENT_LIST_DIR}/color-sensor.h"
 
         "${CMAKE_CURRENT_LIST_DIR}/algo.h"
         "${CMAKE_CURRENT_LIST_DIR}/api.h"
@@ -135,4 +138,6 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/max-usable-range-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/debug-stream-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/serialized-utilities.h"
+        "${CMAKE_CURRENT_LIST_DIR}/frame.h"
+        "${CMAKE_CURRENT_LIST_DIR}/composite-frame.h"
 )
