+++ src/core/streaming.h
@@ -4,6 +4,9 @@
 
 #include "options.h"
 #include "../types.h"
+#include "../depth-sensor.h"
+#include "../color-sensor.h"
+#include "../composite-frame.h"
 #include "info.h"
 #include <functional>
 
@@ -18,14 +21,6 @@ namespace librealsense
     class archive_interface;
     class device_interface;
     class processing_block_interface;
-
-    class sensor_part
-    {
-    public:
-        virtual std::shared_ptr<sensor_interface> get_sensor() const = 0;
-        virtual void set_sensor(std::shared_ptr<sensor_interface> s) = 0;
-        virtual ~sensor_part() = default;
-    };
 
     class context;
 
@@ -79,91 +74,10 @@ namespace librealsense
         virtual void set_c_wrapper(rs2_stream_profile* wrapper) = 0;
     };
 
-    class frame_interface : public sensor_part
-    {
-    public:
-        virtual rs2_metadata_type get_frame_metadata(const rs2_frame_metadata_value& frame_metadata) const = 0;
-        virtual bool supports_frame_metadata(const rs2_frame_metadata_value& frame_metadata) const = 0;
-        virtual int get_frame_data_size() const = 0;
-        virtual const byte* get_frame_data() const = 0;
-        virtual rs2_time_t get_frame_timestamp() const = 0;
-        virtual rs2_timestamp_domain get_frame_timestamp_domain() const = 0;
-        virtual void set_timestamp(double new_ts) = 0;
-        virtual unsigned long long get_frame_number() const = 0;
-
-        virtual void set_timestamp_domain(rs2_timestamp_domain timestamp_domain) = 0;
-        virtual rs2_time_t get_frame_system_time() const = 0;
-        virtual std::shared_ptr<stream_profile_interface> get_stream() const = 0;
-        virtual void set_stream(std::shared_ptr<stream_profile_interface> sp) = 0;
-
-        virtual rs2_time_t get_frame_callback_start_time_point() const = 0;
-        virtual void update_frame_callback_start_ts(rs2_time_t ts) = 0;
-
-        virtual void acquire() = 0;
-        virtual void release() = 0;
-        virtual frame_interface* publish(std::shared_ptr<archive_interface> new_owner) = 0;
-        virtual void unpublish() = 0;
-        virtual void attach_continuation(frame_continuation&& continuation) = 0;
-        virtual void disable_continuation() = 0;
-
-        virtual void log_callback_start(rs2_time_t timestamp) = 0;
-        virtual void log_callback_end(rs2_time_t timestamp) const = 0;
-
-        virtual archive_interface* get_owner() const = 0;
-
-        virtual void mark_fixed() = 0;
-        virtual bool is_fixed() const = 0;
-        virtual void set_blocking(bool state) = 0;
-        virtual bool is_blocking() const = 0;
-
-        virtual void keep() = 0;
-
-        virtual ~frame_interface() = default;
-    };
-
-    struct LRS_EXTENSION_API frame_holder
-    {
-        frame_interface* frame;
-
-        frame_interface* operator->() const
-        {
-            return frame;
-        }
-
-        operator bool() const { return frame != nullptr; }
-
-        operator frame_interface*() const { return frame; }
-
-        frame_holder(frame_interface* f)
-        {
-            frame = f;
-        }
-
-        ~frame_holder();
-
-        frame_holder(frame_holder&& other)
-            : frame(other.frame)
-        {
-            other.frame = nullptr;
-        }
-
-        frame_holder() : frame(nullptr) {}
-
-
-        frame_holder& operator=(frame_holder&& other);
-
-        frame_holder clone() const;
-
-        bool is_blocking() const { return frame->is_blocking(); };
-
-    private:
-        frame_holder& operator=(const frame_holder& other) = delete;
-        frame_holder(const frame_holder& other);
-    };
-
     using on_frame = std::function<void(frame_interface*)>;
     using stream_profiles = std::vector<std::shared_ptr<stream_profile_interface>>;
     using processing_blocks = std::vector<std::shared_ptr<processing_block_interface>>;
+
 
     inline std::ostream& operator << (std::ostream& os, const stream_profiles& profiles)
     {
@@ -174,11 +88,52 @@ namespace librealsense
         return os;
     }
 
-    std::string frame_holder_to_string(const frame_holder & f);
 
-    std::string frame_to_string(const frame_interface & f);
+    // Declaration for solving cyclic usage between frame_to_string and operator<<
+    inline std::ostream& operator<<(std::ostream& out, const frame_interface& f);
 
-    inline std::ostream& operator<<(std::ostream& out, const frame_interface & f)
+    // We place this functions inside an anonymous namespace to avoid error of multiple symbols found on linker.
+    namespace
+    {
+        std::string frame_to_string(const frame_interface& f)
+        {
+            std::ostringstream s;
+
+            if (!&f)
+            {
+                s << "[null]";
+            }
+            else
+            {
+                auto composite = dynamic_cast<const composite_frame*>(&f);
+                if (composite)
+                {
+                    s << "[";
+                    for (int i = 0; i < composite->get_embedded_frames_count(); i++)
+                    {
+                        s << *composite->get_frame(i);
+                    }
+                    s << "]";
+                }
+                else
+                {
+                    s << "[" << rs2_stream_to_string(f.get_stream()->get_stream_type());
+                    s << "/" << f.get_stream()->get_unique_id();
+                    s << " #" << f.get_frame_number();
+                    s << " @" << std::fixed << (double)f.get_frame_timestamp();
+                    s << "]";
+                }
+            }
+            return s.str();
+        }
+
+        std::string frame_holder_to_string(const frame_holder& f)
+        {
+            return frame_to_string(*f.frame);
+        }
+    }
+
+    inline std::ostream& operator<<(std::ostream& out, const frame_interface& f)
     {
         return out << frame_to_string(f);
     }
@@ -283,120 +238,5 @@ namespace librealsense
         virtual bool compress_while_record() const = 0;
 
         virtual bool contradicts(const stream_profile_interface* a, const std::vector<stream_profile>& others) const = 0;
-    };
-
-    class depth_stereo_sensor;
-
-    class color_sensor : public recordable<color_sensor>
-    {
-    public:
-        virtual ~color_sensor() = default;
-
-        void create_snapshot(std::shared_ptr<color_sensor>& snapshot) const override;
-        void enable_recording(std::function<void(const color_sensor&)> recording_function) override {};
-    };
-
-    MAP_EXTENSION(RS2_EXTENSION_COLOR_SENSOR, librealsense::color_sensor);
-
-    class color_sensor_snapshot : public virtual color_sensor, public extension_snapshot
-    {
-    public:
-        color_sensor_snapshot() {}
-
-        void update(std::shared_ptr<extension_snapshot> ext) override
-        {
-        }
-
-        void create_snapshot(std::shared_ptr<color_sensor>& snapshot) const  override
-        {
-            snapshot = std::make_shared<color_sensor_snapshot>(*this);
-        }
-        void enable_recording(std::function<void(const color_sensor&)> recording_function) override
-        {
-            //empty
-        }
-    };
-
-
-    class depth_sensor : public recordable<depth_sensor>
-    {
-    public:
-        virtual float get_depth_scale() const = 0;
-        virtual ~depth_sensor() = default;
-    };
-
-    MAP_EXTENSION(RS2_EXTENSION_DEPTH_SENSOR, librealsense::depth_sensor);
-
-    class depth_sensor_snapshot : public virtual depth_sensor, public extension_snapshot
-    {
-    public:
-        depth_sensor_snapshot(float depth_units) : m_depth_units(depth_units) {}
-        float get_depth_scale() const override
-        {
-            return m_depth_units;
-        }
-
-        void update(std::shared_ptr<extension_snapshot> ext) override
-        {
-            if (auto api = As<depth_sensor>(ext))
-            {
-                m_depth_units = api->get_depth_scale();
-            }
-        }
-        void create_snapshot(std::shared_ptr<depth_sensor>& snapshot) const  override
-        {
-            snapshot = std::make_shared<depth_sensor_snapshot>(*this);
-        }
-        void enable_recording(std::function<void(const depth_sensor&)> recording_function) override
-        {
-            //empty
-        }
-
-    protected:
-        float m_depth_units;
-    };
-
-    class depth_stereo_sensor : public virtual depth_sensor, public recordable<depth_stereo_sensor>
-    {
-    public:
-        virtual float get_stereo_baseline_mm() const = 0;
-    };
-
-    MAP_EXTENSION(RS2_EXTENSION_DEPTH_STEREO_SENSOR, librealsense::depth_stereo_sensor);
-
-    class depth_stereo_sensor_snapshot : public depth_stereo_sensor, public depth_sensor_snapshot
-    {
-    public:
-        depth_stereo_sensor_snapshot(float depth_units, float stereo_bl_mm) :
-            depth_sensor_snapshot(depth_units),
-            m_stereo_baseline_mm(stereo_bl_mm) {}
-
-        float get_stereo_baseline_mm() const override
-        {
-            return m_stereo_baseline_mm;
-        }
-
-        void update(std::shared_ptr<extension_snapshot> ext) override
-        {
-            depth_sensor_snapshot::update(ext);
-
-            if (auto api = As<depth_stereo_sensor>(ext))
-            {
-                m_stereo_baseline_mm = api->get_stereo_baseline_mm();
-            }
-        }
-
-        void create_snapshot(std::shared_ptr<depth_stereo_sensor>& snapshot) const override
-        {
-            snapshot = std::make_shared<depth_stereo_sensor_snapshot>(*this);
-        }
-
-        void enable_recording(std::function<void(const depth_stereo_sensor&)> recording_function) override
-        {
-            //empty
-        }
-
-    private:
-        float m_stereo_baseline_mm;
     };
 }
